+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.1.0
+* Updated to version 84.0.522.63
+* Improved Mac contributing workflow - [PR #174](https://github.com/microsoft/vscode-edge-devtools/pull/174)
+* Webview resources fix to work with newer versions of vscode - [PR #176](https://github.com/microsoft/vscode-edge-devtools/pull/176)
+* Debug mode removed - [PR #183](https://github.com/microsoft/vscode-edge-devtools/pull/183)
+* Introduced headless browser mode - [PR #185](https://github.com/microsoft/vscode-edge-devtools/pull/185)
+
 ## 1.0.9
 * Updated to version 83.0.478.45
 * Added extension settings toggle to toggle network panel on and off - [PR #136](https://github.com/microsoft/vscode-edge-devtools/pull/136)
@@ -8,7 +15,7 @@
 ## 1.0.8
 * Downgraded Edge DevTools version to 81.0.416 (Microsoft Edge Stable)
 * Merged Network Tool into the extension - [PR #128](https://github.com/microsoft/vscode-edge-devtools/pull/128)
-* Renaming tool to "Microsoft Edge (Chromium) Tools for VSCode"
+* Renaming tool to "Microsoft Edge Tools for VS Code"
 
 ## 1.0.7
 * Updated to version 82.0.423